+++ src/crafting/utility/Utility.java
@@ -1,7 +1,7 @@
 package crafting.utility;
 
 import crafting.PoECraftingAssistant;
-import crafting.Settings;
+import crafting.persistence.Settings;
 import crafting.UI.Main;
 import java.awt.AWTException;
 import java.awt.Robot;
